modules/lte_ext: Change usrsock interface with nuttx update
The usrsock interface is changed with nuttx update.

cbe9352f41 net/usrsock: remove the connections limit
Replace xid type to uint64_t to avoid limitation of connections.

+++ sdk/modules/lte_ext/lapiext/alt1250_ioctl_extend.c
@@ -84,7 +84,7 @@ static int postproc_freset_command(FAR struct alt1250_s *dev,
                                    FAR struct alt_container_s *reply,
                                    FAR struct usock_s *usock,
                                    FAR int32_t *usock_result,
-                                   FAR uint8_t *usock_xid,
+                                   FAR uint64_t *usock_xid,
                                    FAR struct usock_ackinfo_s *ackinfo,
                                    unsigned long arg)
 {
@@ -153,7 +153,7 @@ static int send_freset_command(FAR struct alt1250_s *dev,
 int usockreq_ioctl_extend(FAR struct alt1250_s *dev,
                           FAR struct usrsock_request_buff_s *req,
                           FAR int32_t *usock_result,
-                          FAR uint8_t *usock_xid,
+                          FAR uint64_t *usock_xid,
                           FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_ioctl_s *request = &req->request.ioctl_req;
